backend/requirements.txt
fastapi
uvicorn
google-genai
python-multipart
python-dotenv
Pillow
